Combine all of the "guts" package into one file.
